Split capistrano integration from configuration
