+++ Gemfile
@@ -5,4 +5,5 @@ gemspec
 group :test do
   gem 'rspec'
   gem 'capistrano-spec'
+  gem 'pry-debugger'
 end
